+++ packages/charts/src/data/dataScatter.d.ts
@@ -1,7 +1,7 @@
 declare const _default: {
-	year: number;
-	value: number;
-	alt: number;
-	group: string;
+    year: number;
+    value: number;
+    alt: number;
+    group: string;
 }[];
 export default _default;
